refactor(ui-logic): extract audio side-effects out of shared components
- Extract muted state + localStorage + toggle logic from SoundToggleButton
  into new hooks/shared/audio/useSoundToggle.ts
- Extract haptic + TTS wrong-answer feedback + shake-animation state from
  QuizAnswerGrid into new hooks/shared/audio/useWrongAnswerFeedback.ts
- Both components are now pure renderers that call their respective hooks
- Export both hooks from hooks/shared/audio/index.ts
