+++ gamesformykids/components/game/quiz/QuizAnswerGrid.tsx
@@ -1,14 +1,11 @@
 'use client';
 
-import { useState, useEffect, type ReactNode } from 'react';
+import { type ReactNode } from 'react';
 import { useQuizGameStore } from '@/lib/stores/quizGameStore';
 import { QUIZ_THEMES, type QuizTheme } from './quizTheme';
 import { answerButtonClass } from '@/lib/quiz/answerButtonClass';
 import { useKeyboardAnswerSelect } from '@/hooks/shared/useKeyboardAnswerSelect';
-import { useHaptic } from '@/hooks/shared/haptic/useHaptic';
-import { speakHebrew } from '@/lib/utils/speech/speaker';
-
-const WRONG_PHRASES = ['כמעט!', 'נסה שוב!', 'לא נורא, תנסה שוב!'] as const;
+import { useWrongAnswerFeedback } from '@/hooks/shared/audio/useWrongAnswerFeedback';
 
 interface Props {
   choices: string[];
@@ -30,26 +27,13 @@ export function QuizAnswerGrid({
   const selected = useQuizGameStore(s => s.selected);
   const t = QUIZ_THEMES[theme];
   const enabled = selected === null;
-  const { wrong: hapticWrong } = useHaptic();
-  const [shakingChoice, setShakingChoice] = useState<string | null>(null);
+  const { shakingChoice } = useWrongAnswerFeedback(selected, correctValue);
 
   const { focusedIdx } = useKeyboardAnswerSelect(
     choices.length,
     (idx) => onSelect(choices[idx]!),
     enabled,
   );
-
-  useEffect(() => {
-    if (selected !== null && selected !== correctValue) {
-      hapticWrong();
-      const phrase = WRONG_PHRASES[Math.floor(Math.random() * WRONG_PHRASES.length)]!;
-      void speakHebrew(phrase);
-      setShakingChoice(selected);
-      const id = setTimeout(() => setShakingChoice(null), 600);
-      return () => clearTimeout(id);
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [selected]);
 
   const gridClass = cols === 1
     ? 'grid grid-cols-1 gap-3 mb-4'
