+++ gamesformykids/hooks/shared/audio/index.ts
@@ -1,1 +1,3 @@
 export { useGameAudio } from './useGameAudio';
+export { useSoundToggle } from './useSoundToggle';
+export { useWrongAnswerFeedback } from './useWrongAnswerFeedback';
